Hook redux and sagas into app, change app structure to account for bootstrapping files
