--- App.js
@@ -1,24 +1,0 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-import Placeholder from './src/welcome/OnboardingCarousel'
-
-export default class App extends React.Component {
-  render() {
-    return (
-      <View style={styles.container}>
-        <Placeholder />
-
-        <Text>Changes you make will automatically reload.</Text>
-        <Text>Shake your phone to open the developer menu.</Text>
-      </View>
-    )
-  }
-}
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-})
