+++ App/App.test.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import App from './index'
+
+import renderer from 'react-test-renderer'
+
+it('renders without crashing', () => {
+  const rendered = renderer.create(<App />).toJSON()
+  expect(rendered).toBeTruthy()
+})
